pyproject.toml
[project]
name = "my-mcp-servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-core>=2.24.2",
    "google-api-python-client>=2.165.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "tzlocal>=5.3.1",
    "azure-identity>=1.16.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-cosmos>=4.6.0",
]
